refactor(RacingGame): 역할에 따른 폴더 위치 변경
