+++ src/main/java/racingcar/controller/Controller.java
@@ -1,14 +1,12 @@
 package racingcar.controller;
 
 import racingcar.domain.Car;
-import racingcar.domain.RacingGame;
+import racingcar.service.RacingGame;
 import racingcar.utils.Utils;
 import racingcar.validator.Validator;
 import racingcar.view.InputView;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class Controller {
