+++ src/test/java/racingcar/service/RacingGameTest.java
@@ -1,16 +1,14 @@
-package racingcar.domain;
+package racingcar.service;
 
-import camp.nextstep.edu.missionutils.Randoms;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.service.RacingGame;
 
-import java.lang.reflect.Method;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+
 class RacingGameTest {
